refactor: Modularize pty.rs into smaller, maintainable modules
Breaking down the monolithic 1139-line pty.rs file into focused modules:

Module Structure:
- pty/client.rs (38 lines) - Client connection and terminal size management
- pty/socket.rs (61 lines) - Unix socket operations and command parsing
- pty/terminal.rs (119 lines) - Terminal state save/restore and raw mode
- pty/io_handler.rs (172 lines) - PTY I/O operations and scrollback buffer
- pty/session_switcher.rs (155 lines) - Interactive session switching logic
- pty/spawn.rs (865 lines) - Core PTY process spawning and attachment
- pty/mod.rs (14 lines) - Module exports for backward compatibility

Benefits:
- Improved code organization and maintainability
- Better separation of concerns
- Faster incremental compilation
- Easier testing of individual components
- Reduced cognitive load for developers

All tests pass and backward compatibility is maintained.
