+++ Cargo.toml
@@ -3,7 +3,7 @@ name = "detached-shell"
 version = "0.1.1"
 readme = "README.md"
 edition = "2021"
-authors = ["Noras <oss@noras.tech>"]
+authors = ["Kerem Noras <kerem@noras.tech>"]
 description = "Noras.tech's minimalist detachable shell solution · zero configuration · not a complex multiplexer, just persistent sessions"
 repository = "https://github.com/NorasTech/detached-shell"
 homepage = "https://github.com/NorasTech/detached-shell"
